+++ Changelog.md
@@ -48,6 +48,7 @@ With the port to Bootstrap 3, app/views/terms/default.haml has a new structure. 
 * Support color themes [#6033](https://github.com/diaspora/diaspora/pull/6033)
 * Add mobile services and privacy settings pages [#6086](https://github.com/diaspora/diaspora/pull/6086)
 * Optionally make your extended profile details public [#6162](https://github.com/diaspora/diaspora/pull/6162)
+* Add admin dashboard showing latest diaspora\* version [#6216](https://github.com/diaspora/diaspora/pull/6216)
 
 # 0.5.3.0